Enables Memcached Session and Cache testing on Travis.
Corrects bugs in construction of Memcached Session and Cache objects which can cause failures if the default configuration is not in use.

Implements Unit Tests of most class methods to ensure that they really are working.

Refactors Unit Tests for Joomla\Cache\Cache to remove calls to methods that are specific to a hard coded class name and replace it with something that subclasses can set via setUp - reduce copy and paste of test code and instead Cache engine tests should inherit from CacheTest

Refactor for Joomla\Session\Storage to do the same in theory.  In practice, because it is an Abstract class the calling mechanisms are not compatible.  Instead of requiring every storage engine to re-implement each and every test, testing methodology was embedded into the new StorageCase interface which every Storage Test Class should implement.  Logic should be usable for any class except for StorageTest, which both implements and overrides almost every method.