Fix the SQLSVR tests
Information based on the references below
It seems that PHPUnit deletes the content of the table in question before setting up the table to execute a test. That only works, if the table already exists. Therefore I’ve added a create table jos_dbtest statement with the PDO to fix the issue

PHP unit will perform a ‘clean insert’ operation on our database using the the dataset returned by getDataSet. The clean insert operation will truncate or delete_all any tables in the data set (not all tables in the database) and then insert the rows in the dataset. The solution is to create the table before running the test.

Additionally we can modify the default functionality by overriding the getSetUpOperation() and getTearDownOperation() methods

- https://andreas.heigl.org/2015/08/16/database-testing-with-phpunit-and-sqlite/
- http://stackoverflow.com/questions/10948273/loading-schema-from-sql-file-in-phpunit-test