Option to output the database query without pre tags
When you are developing and debugging with XDebug and want to get the query, you can do a `$query->dump()` but this includes the pre tags. You need to copy paste the query, remove the tags and then you can execute the query. With this change you can easily do `$query->dump(false)` to get the query without pre tags and just copy paste and execute.